Deleted application.php since not final
I made a mistake and included application.php in master. I'd like to keep all files that are NOT final off the master branch.

For everyone's information, if ever you want to make a new page and it's a work in progress, please use new branch. It's good practice.

When the new page is finished and ready to go live, merge into master.

Apologies regarding my mistake.